+++ src/secops/client.py
@@ -14,9 +14,11 @@
 
 """Main client for Google SecOps SDK."""
 
-from typing import Optional, Dict, Any
+from typing import Any, Dict, Optional, Union
+
 from google.auth.credentials import Credentials
-from secops.auth import SecOpsAuth
+
+from secops.auth import RetryConfig, SecOpsAuth
 from secops.chronicle import ChronicleClient
 
 
@@ -29,6 +31,7 @@ class SecOpsClient:
         service_account_path: Optional[str] = None,
         service_account_info: Optional[Dict[str, Any]] = None,
         impersonate_service_account: Optional[str] = None,
+        retry_config: Optional[Union[RetryConfig, Dict[str, Any], bool]] = None,
     ):
         """Initialize the SecOps client.
 
@@ -37,12 +40,15 @@ class SecOpsClient:
             service_account_path: Optional path to service account JSON key file
             service_account_info: Optional service account JSON key data as dict
             impersonate_service_account: Optional service account to impersonate
+            retry_config: Request retry configurations.
+                If set to false, retry will be disabled.
         """
         self.auth = SecOpsAuth(
             credentials=credentials,
             service_account_path=service_account_path,
             service_account_info=service_account_info,
             impersonate_service_account=impersonate_service_account,
+            retry_config=retry_config,
         )
         self._chronicle = None