+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2025-09-11
+### Added
+- Default retry mechanism for all SecOps requests
+## Updated
+- Clients (SecOpsClient, ChronicleClient) to accept user define retry configuration
+
 ## [0.16.0] - 2025-09-10
 ### Added
 - Support for import native dashboard method